+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -163,9 +163,9 @@ Changes in existing checks
   <clang-tidy/checks/bugprone/reserved-identifier>`, so that it does not warn
   on macros starting with underscore and lowercase letter.
 
-- Improved :doc:`modernize-loop-convert <clang-tidy/checks/modernize/loop-convert>` to
-  refactor container based for loops that initialize iterators with free function calls
-  to ``begin``, ``end``, or ``size``.
+- Improved :doc:`modernize-loop-convert
+  <clang-tidy/checks/modernize/loop-convert>` to support for-loops with
+  iterators initialized by free functions like `begin`, `end`, or `size`.
 
 Removed checks
 ^^^^^^^^^^^^^^